+++ xcube/core/gen/gen.py
@@ -217,7 +217,7 @@ def _process_input(input_processor: InputProcessor,
 
     # noinspection PyShadowingNames
     def step1(input_slice):
-        return input_processor.pre_process(input_slice)
+        return input_processor.pre_process(input_slice, output_region=output_region)
 
     steps.append((step1, 'pre-processing input slice'))
 
@@ -238,8 +238,8 @@ def _process_input(input_processor: InputProcessor,
 
     # noinspection PyShadowingNames
     def step4(input_slice):
-        if output_region:
-            input_slice = input_processor.get_spatial_subset(input_slice, output_region)
+        # if output_region:
+        #     input_slice = input_processor.get_spatial_subset(input_slice, output_region)
         return input_processor.process(input_slice,
                                        dst_size=output_size,
                                        dst_region=output_region,